python_src/requirements.txt
Flask==0.10.1
python-networkmanager=2.0.1